requirements.txt
absl-py==2.1.0
altair==5.3.0
app-store-scraper==0.3.5
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
Flask==3.0.3
flask-ngrok==0.0.25
flatbuffers==24.3.25
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
google-play-scraper==1.2.6
grpcio==1.63.0
gspread==6.1.2
h5py==3.11.0
httplib2==0.22.0
idna==2.10
install==1.3.5
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mysql-connector-python==9.0.0
namex==0.0.8
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydeck==0.9.1
Pygments==2.18.0
pyodbc==5.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.1.0
referencing==0.35.1
regex==2024.4.16
requests==2.23.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.34.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.25.11
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
